feature: update ocean current processing to consume gsla_data with direction speed and gsla values
